+++ text2qti/xml_assessment.py
@@ -85,7 +85,7 @@ END_ITEM = '''\
 '''
 
 
-ITEM_METADATA_MCTF_MULTANS_NUM = '''\
+ITEM_METADATA_MCTF_SHORTANS_MULTANS_NUM = '''\
         <itemmetadata>
           <qtimetadata>
             <qtimetadatafield>
@@ -108,7 +108,7 @@ ITEM_METADATA_MCTF_MULTANS_NUM = '''\
         </itemmetadata>
 '''
 
-ITEM_METADATA_ESSAY = ITEM_METADATA_MCTF_MULTANS_NUM.replace('{original_answer_ids}', '')
+ITEM_METADATA_ESSAY = ITEM_METADATA_MCTF_SHORTANS_MULTANS_NUM.replace('{original_answer_ids}', '')
 
 ITEM_METADATA_UPLOAD = ITEM_METADATA_ESSAY
 
@@ -135,6 +135,19 @@ ITEM_PRESENTATION_MCTF_CHOICE = '''\
 ITEM_PRESENTATION_MULTANS = ITEM_PRESENTATION_MCTF.replace('Single', 'Multiple')
 
 ITEM_PRESENTATION_MULTANS_CHOICE = ITEM_PRESENTATION_MCTF_CHOICE
+
+ITEM_PRESENTATION_SHORTANS = '''\
+        <presentation>
+          <material>
+            <mattext texttype="text/html">{question_html_xml}</mattext>
+          </material>
+          <response_str ident="response1" rcardinality="Single">
+            <render_fib>
+              <response_label ident="answer1" rshuffle="No"/>
+            </render_fib>
+          </response_str>
+        </presentation>
+'''
 
 ITEM_PRESENTATION_ESSAY = '''\
         <presentation>
@@ -224,6 +237,41 @@ ITEM_RESPROCESSING_MCTF_INCORRECT_FEEDBACK = '''\
           </respcondition>
 '''
 
+ITEM_RESPROCESSING_SHORTANS_GENERAL_FEEDBACK = ITEM_RESPROCESSING_MCTF_GENERAL_FEEDBACK
+
+ITEM_RESPROCESSING_SHORTANS_CHOICE_FEEDBACK = '''\
+          <respcondition continue="Yes">
+            <conditionvar>
+              <varequal respident="response1">{answer_xml}</varequal>
+            </conditionvar>
+            <displayfeedback feedbacktype="Response" linkrefid="{ident}_fb"/>
+          </respcondition>
+'''
+
+ITEM_RESPROCESSING_SHORTANS_SET_CORRECT_WITH_FEEDBACK = '''\
+          <respcondition continue="No">
+            <conditionvar>
+{varequal}
+            </conditionvar>
+            <setvar action="Set" varname="SCORE">100</setvar>
+            <displayfeedback feedbacktype="Response" linkrefid="correct_fb"/>
+          </respcondition>
+'''
+
+ITEM_RESPROCESSING_SHORTANS_SET_CORRECT_NO_FEEDBACK = '''\
+          <respcondition continue="No">
+            <conditionvar>
+{varequal}
+            </conditionvar>
+            <setvar action="Set" varname="SCORE">100</setvar>
+          </respcondition>
+'''
+
+ITEM_RESPROCESSING_SHORTANS_SET_CORRECT_VAREQUAL = '''\
+              <varequal respident="response1">{answer_xml}</varequal>'''
+
+ITEM_RESPROCESSING_SHORTANS_INCORRECT_FEEDBACK = ITEM_RESPROCESSING_MCTF_INCORRECT_FEEDBACK
+
 ITEM_RESPROCESSING_MULTANS_GENERAL_FEEDBACK = ITEM_RESPROCESSING_MCTF_GENERAL_FEEDBACK
 
 ITEM_RESPROCESSING_MULTANS_CHOICE_FEEDBACK = ITEM_RESPROCESSING_MCTF_CHOICE_FEEDBACK
@@ -252,14 +300,12 @@ ITEM_RESPROCESSING_MULTANS_SET_CORRECT_NO_FEEDBACK = '''\
 '''
 
 ITEM_RESPROCESSING_MULTANS_SET_CORRECT_VAREQUAL_CORRECT = '''\
-                <varequal respident="response1">{ident}</varequal>
-'''
+                <varequal respident="response1">{ident}</varequal>'''
 
 ITEM_RESPROCESSING_MULTANS_SET_CORRECT_VAREQUAL_INCORRECT = '''\
                 <not>
                   <varequal respident="response1">{ident}</varequal>
-                </not>
-'''
+                </not>'''
 
 ITEM_RESPROCESSING_MULTANS_INCORRECT_FEEDBACK = ITEM_RESPROCESSING_MCTF_INCORRECT_FEEDBACK
 
@@ -338,7 +384,7 @@ ITEM_RESPROCESSING_END = '''\
 '''
 
 
-ITEM_FEEDBACK_MCTF_MULTANS_NUM_GENERAL = '''\
+ITEM_FEEDBACK_MCTF_SHORTANS_MULTANS_NUM_GENERAL = '''\
         <itemfeedback ident="general_fb">
           <flow_mat>
             <material>
@@ -348,7 +394,7 @@ ITEM_FEEDBACK_MCTF_MULTANS_NUM_GENERAL = '''\
         </itemfeedback>
 '''
 
-ITEM_FEEDBACK_MCTF_MULTANS_NUM_CORRECT = '''\
+ITEM_FEEDBACK_MCTF_SHORTANS_MULTANS_NUM_CORRECT = '''\
         <itemfeedback ident="correct_fb">
           <flow_mat>
             <material>
@@ -358,7 +404,7 @@ ITEM_FEEDBACK_MCTF_MULTANS_NUM_CORRECT = '''\
         </itemfeedback>
 '''
 
-ITEM_FEEDBACK_MCTF_MULTANS_NUM_INCORRECT = '''\
+ITEM_FEEDBACK_MCTF_SHORTANS_MULTANS_NUM_INCORRECT = '''\
         <itemfeedback ident="general_incorrect_fb">
           <flow_mat>
             <material>
@@ -368,7 +414,7 @@ ITEM_FEEDBACK_MCTF_MULTANS_NUM_INCORRECT = '''\
         </itemfeedback>
 '''
 
-ITEM_FEEDBACK_MCTF_MULTANS_NUM_INDIVIDUAL = '''\
+ITEM_FEEDBACK_MCTF_SHORTANS_MULTANS_NUM_INDIVIDUAL = '''\
         <itemfeedback ident="{ident}_fb">
           <flow_mat>
             <material>
@@ -411,11 +457,12 @@ def assessment(*, quiz: Quiz, assessment_identifier: str, title_xml: str) -> str
         xml.append(START_ITEM.format(question_identifier=f'text2qti_question_{question.id}',
                                      question_title=question.title_xml))
 
-        if question.type in ('true_false_question', 'multiple_choice_question', 'multiple_answers_question'):
-            item_metadata = ITEM_METADATA_MCTF_MULTANS_NUM
+        if question.type in ('true_false_question', 'multiple_choice_question',
+                             'short_answer_question', 'multiple_answers_question'):
+            item_metadata = ITEM_METADATA_MCTF_SHORTANS_MULTANS_NUM
             original_answer_ids = ','.join(f'text2qti_choice_{c.id}' for c in question.choices)
         elif question.type == 'numerical_question':
-            item_metadata = ITEM_METADATA_MCTF_MULTANS_NUM
+            item_metadata = ITEM_METADATA_MCTF_SHORTANS_MULTANS_NUM
             original_answer_ids = f'text2qti_numerical_{question.id}'
         elif question.type == 'essay_question':
             item_metadata = ITEM_METADATA_ESSAY
@@ -442,6 +489,8 @@ def assessment(*, quiz: Quiz, assessment_identifier: str, title_xml: str) -> str
             choices = '\n'.join(item_presentation_choice.format(ident=f'text2qti_choice_{c.id}', choice_html_xml=c.choice_html_xml)
                                                                 for c in question.choices)
             xml.append(item_presentation.format(question_html_xml=question.question_html_xml, choices=choices))
+        elif question.type == 'short_answer_question':
+            xml.append(ITEM_PRESENTATION_SHORTANS.format(question_html_xml=question.question_html_xml))
         elif question.type == 'numerical_question':
             xml.append(ITEM_PRESENTATION_NUM.format(question_html_xml=question.question_html_xml))
         elif question.type == 'essay_question':
@@ -474,6 +523,25 @@ def assessment(*, quiz: Quiz, assessment_identifier: str, title_xml: str) -> str
                 resprocessing.append(ITEM_RESPROCESSING_MCTF_INCORRECT_FEEDBACK)
             resprocessing.append(ITEM_RESPROCESSING_END)
             xml.extend(resprocessing)
+        elif question.type == 'short_answer_question':
+            resprocessing = []
+            resprocessing.append(ITEM_RESPROCESSING_START)
+            if question.feedback_raw is not None:
+                resprocessing.append(ITEM_RESPROCESSING_SHORTANS_GENERAL_FEEDBACK)
+            for choice in question.choices:
+                if choice.feedback_raw is not None:
+                    resprocessing.append(ITEM_RESPROCESSING_SHORTANS_CHOICE_FEEDBACK.format(ident=f'text2qti_choice_{choice.id}', answer_xml=choice.choice_xml))
+            varequal = []
+            for choice in question.choices:
+                varequal.append(ITEM_RESPROCESSING_SHORTANS_SET_CORRECT_VAREQUAL.format(answer_xml=choice.choice_xml))
+            if question.correct_feedback_raw is not None:
+                resprocessing.append(ITEM_RESPROCESSING_SHORTANS_SET_CORRECT_WITH_FEEDBACK.format(varequal='\n'.join(varequal)))
+            else:
+                resprocessing.append(ITEM_RESPROCESSING_SHORTANS_SET_CORRECT_NO_FEEDBACK.format(varequal='\n'.join(varequal)))
+            if question.incorrect_feedback_raw is not None:
+                resprocessing.append(ITEM_RESPROCESSING_SHORTANS_INCORRECT_FEEDBACK)
+            resprocessing.append(ITEM_RESPROCESSING_END)
+            xml.extend(resprocessing)
         elif question.type == 'multiple_answers_question':
             resprocessing = []
             resprocessing.append(ITEM_RESPROCESSING_START)
@@ -489,9 +557,9 @@ def assessment(*, quiz: Quiz, assessment_identifier: str, title_xml: str) -> str
                 else:
                     varequal.append(ITEM_RESPROCESSING_MULTANS_SET_CORRECT_VAREQUAL_INCORRECT.format(ident=f'text2qti_choice_{choice.id}'))
             if question.correct_feedback_raw is not None:
-                resprocessing.append(ITEM_RESPROCESSING_MULTANS_SET_CORRECT_WITH_FEEDBACK.format(varequal=''.join(varequal)))
+                resprocessing.append(ITEM_RESPROCESSING_MULTANS_SET_CORRECT_WITH_FEEDBACK.format(varequal='\n'.join(varequal)))
             else:
-                resprocessing.append(ITEM_RESPROCESSING_MULTANS_SET_CORRECT_NO_FEEDBACK.format(varequal=''.join(varequal)))
+                resprocessing.append(ITEM_RESPROCESSING_MULTANS_SET_CORRECT_NO_FEEDBACK.format(varequal='\n'.join(varequal)))
             if question.incorrect_feedback_raw is not None:
                 resprocessing.append(ITEM_RESPROCESSING_MULTANS_INCORRECT_FEEDBACK)
             resprocessing.append(ITEM_RESPROCESSING_END)
@@ -530,19 +598,21 @@ def assessment(*, quiz: Quiz, assessment_identifier: str, title_xml: str) -> str
         else:
             raise ValueError
 
-        if question.type in ('true_false_question', 'multiple_choice_question', 'multiple_answers_question',
+        if question.type in ('true_false_question', 'multiple_choice_question',
+                             'short_answer_question', 'multiple_answers_question',
                              'numerical_question', 'essay_question', 'file_upload_question'):
             if question.feedback_raw is not None:
-                xml.append(ITEM_FEEDBACK_MCTF_MULTANS_NUM_GENERAL.format(feedback=question.feedback_html_xml))
+                xml.append(ITEM_FEEDBACK_MCTF_SHORTANS_MULTANS_NUM_GENERAL.format(feedback=question.feedback_html_xml))
             if question.correct_feedback_raw is not None:
-                xml.append(ITEM_FEEDBACK_MCTF_MULTANS_NUM_CORRECT.format(feedback=question.correct_feedback_html_xml))
+                xml.append(ITEM_FEEDBACK_MCTF_SHORTANS_MULTANS_NUM_CORRECT.format(feedback=question.correct_feedback_html_xml))
             if question.incorrect_feedback_raw is not None:
-                xml.append(ITEM_FEEDBACK_MCTF_MULTANS_NUM_INCORRECT.format(feedback=question.incorrect_feedback_html_xml))
-        if question.type in ('true_false_question', 'multiple_choice_question', 'multiple_answers_question'):
+                xml.append(ITEM_FEEDBACK_MCTF_SHORTANS_MULTANS_NUM_INCORRECT.format(feedback=question.incorrect_feedback_html_xml))
+        if question.type in ('true_false_question', 'multiple_choice_question',
+                             'short_answer_question', 'multiple_answers_question'):
             for choice in question.choices:
                 if choice.feedback_raw is not None:
-                    xml.append(ITEM_FEEDBACK_MCTF_MULTANS_NUM_INDIVIDUAL.format(ident=f'text2qti_choice_{choice.id}',
-                                                                                feedback=choice.feedback_html_xml))
+                    xml.append(ITEM_FEEDBACK_MCTF_SHORTANS_MULTANS_NUM_INDIVIDUAL.format(ident=f'text2qti_choice_{choice.id}',
+                                                                                         feedback=choice.feedback_html_xml))
 
         xml.append(END_ITEM)